+++ test/browser-cjs/index.spec.cjs
@@ -2,13 +2,13 @@ const { expect } = require("chai");
 const sinon = require("sinon");
 const {
     decodeTime,
-    encodeTime,
     detectPRNG,
-    encodeRandom,
+    encodeTime,
+    fixULIDBase32,
+    isValid,
     monotonicFactory,
-    randomChar,
     ulid
-} = require("../dist/ulid.js");
+} = require("../../dist/browser/index.cjs");
 
 describe("ulid", function() {
     describe("decodeTime", function() {
@@ -16,6 +16,16 @@ describe("ulid", function() {
             const timestamp = Date.now();
             const id = ulid(timestamp);
             expect(decodeTime(id)).to.equal(timestamp);
+        });
+
+        it("should return correct timestamp for README example", function() {
+            const id = "01ARYZ6S41TSV4RRFFQ69G5FAV";
+            expect(decodeTime(id)).to.equal(1469918176385);
+        });
+
+        it("should support decoding lower-case ULIDs", function() {
+            const id = "01aryz6s41tsv4rrffq69g5fav";
+            expect(decodeTime(id)).to.equal(1469918176385);
         });
 
         it("should accept the maximum allowed timestamp", function() {
@@ -55,16 +65,6 @@ describe("ulid", function() {
             it("should be between 0 and 1", function() {
                 expect(this.prng()).to.satisfy(num => num >= 0 && num <= 1);
             });
-        });
-    });
-
-    describe("encodeRandom", function() {
-        beforeEach(function() {
-            this.prng = detectPRNG();
-        });
-
-        it("should return correct length", function() {
-            expect(encodeRandom(12, this.prng)).to.have.a.lengthOf(12);
         });
     });
 
@@ -114,6 +114,54 @@ describe("ulid", function() {
         });
     });
 
+    describe("fixULIDBase32", function() {
+        it("should return the same ULID if no typos or hyphens are present", function() {
+            expect(fixULIDBase32("01ARYZ6S41TSV4RRFFQ69G5FAV")).to.equal(
+                "01ARYZ6S41TSV4RRFFQ69G5FAV"
+            );
+        });
+
+        it("should return the correct ULID with typos fixed", function() {
+            expect(fixULIDBase32("oLARYZ6S41TSV4RRFFQ69G5FAV")).to.equal(
+                "01ARYZ6S41TSV4RRFFQ69G5FAV"
+            );
+        });
+
+        it("should return the correct ULID with hyphens removed", function() {
+            expect(fixULIDBase32("01ARYZ6-S41TSV4RRF-FQ69G5FAV")).to.equal(
+                "01ARYZ6S41TSV4RRFFQ69G5FAV"
+            );
+        });
+
+        it("should return the correct ULID with typos fixed and hyphens removed", function() {
+            expect(fixULIDBase32("oLARYZ6-S41TSV4RRF-FQ69G5FAV")).to.equal(
+                "01ARYZ6S41TSV4RRFFQ69G5FAV"
+            );
+        });
+    });
+
+    describe("isValid", function() {
+        it("should return true for valid ULIDs (uppercase)", function() {
+            expect(isValid("01ARYZ6S41TSV4RRFFQ69G5FAV")).to.be.true;
+        });
+
+        it("should return true for valid ULIDs (lowercase)", function() {
+            expect(isValid("01aryz6s41tsv4rrffq69g5fav")).to.be.true;
+        });
+
+        it("should return false for invalid ULIDs (wrong length)", function() {
+            expect(isValid("01ARYZ6S41TSV4RRFFQ69G5FA")).to.be.false;
+        });
+
+        it("should return false for invalid ULIDs (wrong characters)", function() {
+            expect(isValid("01ARYZ6S41TSV4RRFFQ69G5FA!")).to.be.false;
+        });
+
+        it("should return false for invalid ULIDs (wrong type)", function() {
+            expect(isValid(123)).to.be.false;
+        });
+    });
+
     describe("monotonicFactory", function() {
         it("outputs a function", function() {
             expect(monotonicFactory()).to.be.a("function");
@@ -153,26 +201,6 @@ describe("ulid", function() {
                     expect(this.ulid()).to.equal("01ARYZ6S41YYYYYYYYYYYYYYZ1");
                 });
             });
-        });
-    });
-
-    describe("randomChar", function() {
-        beforeEach(function() {
-            this.prng = detectPRNG();
-        });
-
-        it("should never return undefined", function() {
-            for (let x = 0; x < 320000; x++) {
-                const randChar = randomChar(this.prng);
-                expect(randChar).to.not.be.undefined;
-            }
-        });
-
-        it("should never return an empty string", function() {
-            for (let x = 0; x < 320000; x++) {
-                const randChar = randomChar(this.prng);
-                expect(randChar).to.not.equal("");
-            }
         });
     });